Updated test_metrics and utility scripts
