+++ utils/dataloaders/__init__.py
@@ -1,3 +1,4 @@
-__all__ = ["USPSDataset0_6"]
+__all__ = ["USPSDataset0_6", "USPSH5_Digit_7_9_Dataset"]
 
 from .usps_0_6 import USPSDataset0_6
+from .uspsh5_7_9 import USPSH5_Digit_7_9_Dataset
