+++ utils/load_metric.py
@@ -3,7 +3,7 @@ import copy
 import numpy as np
 import torch.nn as nn
 
-from .metrics import EntropyPrediction
+from .metrics import EntropyPrediction, F1Score
 
 
 class MetricWrapper(nn.Module):
@@ -35,7 +35,7 @@ class MetricWrapper(nn.Module):
             case "entropy":
                 return EntropyPrediction()
             case "f1":
-                raise NotImplementedError("F1 score not implemented yet")
+                raise F1Score()
             case "recall":
                 raise NotImplementedError("Recall score not implemented yet")
             case "precision":
